fix(tauri): stamp packaged sidecar versions
Build the Tauri-packaged hecate and hecate-acp sidecars with the same version that the native app bundle uses instead of leaving internal/version.Version at its local-build default of dev.

Add a dedicated tauri-build-sidecars recipe that resolves TAURI_VERSION, the current git tag, or the Tauri Cargo.toml version, then injects it through Go ldflags for both sidecar binaries. The Tauri matrix now calls that recipe with the release tag before staging platform-specific externalBin names.

Update the Tauri agent guidance so future native-app changes use the versioned sidecar path and understand why plain just build is not sufficient for packaged bundles.

Verified with TAURI_VERSION=v1.2.3-alpha.4 just tauri-build-sidecars, then ./hecate --version and ./hecate-acp --version both reporting 1.2.3-alpha.4.
